# perf: calculate curve length in single iteration