--- a/R/fwise.R
+++ b/R/fwise.R
@@ -97,8 +97,8 @@ tf_fmean <- function(x, arg = tf_arg(x)) {
   assert_arg(arg = arg, x = x)
   x_ <- tf_interpolate(x, arg = arg)
   arg <- ensure_list(arg)
-  length <- map_dbl(arg, \(x) max(x) - min(x))
-  tf_integrate(x_) / length
+  len <- map_dbl(arg, \(x) max(x) - min(x))
+  tf_integrate(x_) / len
 }
 
 #' @export
@@ -108,11 +108,11 @@ tf_fvar <- function(x, arg = tf_arg(x)) {
   assert_tf(x)
   assert_arg(arg = arg, x = x)
   arg <- ensure_list(arg)
-  length <- map_dbl(arg, \(x) max(x) - min(x))
+  len <- map_dbl(arg, \(x) max(x) - min(x))
   x_ <- tfd(x, arg = arg) # cast to tfd to avoid repeatedly casting back to tfb
   x_mean <- tf_interpolate(x, arg = arg) |> tf_fmean()
   x_c <- x_ - x_mean
-  tf_integrate(x_c^2) / length
+  tf_integrate(x_c^2) / len
 }
 
 #' @export
@@ -135,7 +135,7 @@ tf_crosscov <- function(x, y, arg = tf_arg(x)) {
   assert_arg(arg = arg, x = x)
   assert_true(identical(tf_domain(x), tf_domain(y)))
   arg <- ensure_list(arg)
-  length <- map_dbl(arg, \(x) max(x) - min(x))
+  len <- map_dbl(arg, \(x) max(x) - min(x))
   # set up common args &
   # cast to tfd to avoid repeatedly casting back to tfb
   x_ <- tfd(x, arg = arg)
@@ -144,7 +144,7 @@ tf_crosscov <- function(x, y, arg = tf_arg(x)) {
   y_mean <- tf_interpolate(y, arg = arg) |> tf_fmean()
   x_c <- x_ - x_mean
   y_c <- y_ - y_mean
-  tf_integrate(x_c * y_c) / length
+  tf_integrate(x_c * y_c) / len
 }
 
 #' @export