use a single iteration for calculating curve length, somewhat suprised the memory went up:

``` r
args <- list(runif(100), runif(100), runif(100), runif(100), runif(100))
bench::mark(
  old = purrr::map_dbl(args, max) - purrr::map_dbl(args, min),
  new = purrr::map_dbl(args, \(x) max(x) - min(x))
)
#> # A tibble: 2 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 old          89.7µs     92µs    10497.    1.55MB     35.4
#> 2 new          12.9µs   13.4µs    73124.    3.35MB     51.2
```

<sup>Created on 2025-03-12 with [reprex v2.1.1](https://reprex.tidyverse.org)</sup>